refactor vegatimeseriesplot again
