+++ love/src/components/UIF/ComponentIndex.jsx
@@ -181,10 +181,10 @@ export const uifIndex = {
     component: require('../GeneralPurpose/TimeSeriesPlot/TimeSeriesPlot.container').default,
     schema: require('../GeneralPurpose/TimeSeriesPlot/TimeSeriesPlot.container').schema,
   },
-  VegaTimeseriesPlot: {
-    component: require('../GeneralPurpose/TimeSeriesPlot/VegaTimeSeriesPlot.container').default,
-    schema: require('../GeneralPurpose/TimeSeriesPlot/VegaTimeSeriesPlot.container').schema,
-  },
+  // VegaTimeseriesPlot: {
+  //   component: require('../GeneralPurpose/VegaTimeseriesPlot/VegaTimeseriesPlot.container').default,
+  //   schema: require('../GeneralPurpose/VegaTimeseriesPlot/VegaTimeseriesPlot.container').schema,
+  // },
   // TelemetryLog: {
   //   component: require('../TelemetryLog/TelemetryLog.container').default,
   //   schema: {
